working with forward ref
