+++ intro-component-app/src/content/content.js
@@ -1,0 +1,19 @@
+import React from "react";
+
+const contentParagraph =
+  "See how experienced developers solve problems in real-time. Watching scripted tutorials is great, but understanding how developers think is invaluable.";
+const heading = "Learn to code by watching others";
+const offer = (
+  <p>
+    <strong>Try it free 7 days</strong> then $20/mo. thereafter.
+  </p>
+);
+
+const termsAndConditions = (
+  <p>
+    By clicking the button, you are agreeing to our{" "}
+    <span>Terms and Services</span>
+  </p>
+);
+
+export { contentParagraph, heading, offer, termsAndConditions };
